+++ src/core/tamper/multiplespaces.py
@@ -17,10 +17,8 @@ import random
 from src.utils import settings
 
 """
-Adds multiple spaces around OS commands
-Notes:
-Useful to bypass very weak and bespoke web application firewalls
-that has poorly written permissive regular expressions
+About: Adds multiple spaces around OS commands
+Notes: Useful to bypass very weak and bespoke web application firewalls that has poorly written permissive regular expressions.
 """
 
 __tamper__ = "multiplespaces"