+++ src/core/tamper/sleep2usleep.py
@@ -17,10 +17,9 @@ import sys
 from src.utils import settings
 
 """
-Replaces "sleep" with "usleep" [1] command in the generated payloads.
-[1] http://man7.org/linux/man-pages/man3/usleep.3.html
-Notes:
-  * This tamper script works against *nix targets.
+About: Replaces "sleep" with "usleep" [1] command in the generated payloads.
+  [1] http://man7.org/linux/man-pages/man3/usleep.3.html
+Notes: This tamper script works against *nix targets.
 """
 
 __tamper__ = "sleep2usleep"