+++ src/core/tamper/space2htab.py
@@ -16,9 +16,8 @@ For more see the file 'readme/COPYING' for copying permission.
 from src.utils import settings
 
 """
-Replaces space character (' ') with horizontal tab ('%09')
-Notes:
-  * This tamper script works against all targets.
+About: Replaces space character (' ') with horizontal tab ('%09')
+Notes: This tamper script works against all targets.
 """
 
 __tamper__ = "space2htab"