+++ src/core/tamper/space2ifs.py
@@ -17,11 +17,10 @@ import sys
 from src.utils import settings
 
 """
-Replaces space character (' ') with the internal field separator ('$IFS').
+About: Replaces space character (' ') with the internal field separator ('$IFS').
 The internal field separator refers to a variable which defines the character 
 or characters used to separate a pattern into tokens for some operations.
-Notes:
-  * This tamper script works against *nix targets.
+Notes: This tamper script works against *nix targets.
 """
 
 __tamper__ = "space2ifs"