+++ src/core/tamper/space2vtab.py
@@ -16,9 +16,8 @@ For more see the file 'readme/COPYING' for copying permission.
 from src.utils import settings
 
 """
-Replaces space character (' ') with vertical tab ('%0b').
-Notes:
-  * This tamper script works against Windows targets.
+About: Replaces space character (' ') with vertical tab ('%0b').
+Notes: This tamper script works against Windows targets.
 """
 
 __tamper__ = "space2vtab"